fix: do not create files for OIDC/SAML settings
With the introduction of the administrative web interface, the service now
modifies the .env file to reflect user changes. The dotenv library does not
deal well with certain characters within a value, such as hash (#), which
may be found in IdP metadata. Writing complex values to separate files will
work, however the service now has to manage creating and removing these
files, which can result in mistakenly deleting the wrong files. This would
be the first time that the service is deleting any files. Trying to escape
the problematic characters (e.g. # becomes XX-HASH-XX) means that when
reading the values from process.env, it will be necessary to perform the
reverse escape.

Previously, the service was creating uniquely named files to store data such
as the SAML IdP metadata or public certificate. Any time that provider was
modified by the user, another uniquely named file was created.

This change now prefers to use AUTH_PROVIDERS if it is defined, and remove
the classic OIDC and SAML settings when writing the configuration changes to
the .env file. As such, all provider data is stored within the JSON blob and
that avoids the issues with dotenv. However, now customers and support will
need to examine a JSON blob when debugging the service.

npm test passes