feat: force authn available per identity provider
Two new settings have been added that relate to the FORCE_AUTHN boolean
setting that has been available for some time.

For OIDC identity providers, the new setting OIDC_MAX_AGE is available for
setting the maximum age (in seconds) of an authenticated session, after
which the user must authenticate. Setting this to 0 is effectively the same
as setting FORCE_AUTHN to true.

For SAML identity providers, the new setting SAML_FORCE_AUTHN is available
to force the user to authenticate with each login request.

Both of these new settings will take affect if FORCE_AUTHN is not set or is
false. That is, if the client application (such as ALM License Server) asks
that the user be forced to authenticate, then the presence or absence of the
new provider settings will be ignored and authentication will be required.
Similarly, if the client application does not set FORCE_AUTHN, then it is up
to the provider configuration to determine if authentication will be
required or not.

npm test passes