fix: clear temporary env repository after refresh
Previously the temporary settings repository was being cleared when the POST
/settings/apply endpoint was invoked. However, that is resulting in the
admin interface getting stale results because the Node.js process is slow to
restart (unlike on macOS where it restarts immediately). Now the temp repo
is cleared when the environment is refreshed, eliminating that particular
timing issue.

npm test passes