fix: invalidate cache data when logging into webui
If after the admin logged in to the web interface the auth token expired,
then the next time the admin authenticated they would get an error. This
happened under certain circumstances, such as refreshing the main page, but
not when invoking logout and logging in again.

The fix is that when the login endpoint is executed, all cached data will be
invalidated and thus Redux-Toolkit will be forced to refetch the data.
Without this, RTK will remember the failed response and replay that.