fix: encode certain provider values for JSON safety
Certain fields in an auth provider definition may give the JSON parser
trouble when reading the settings. One specific example is to paste the XML
metadata for the Azure AD SAML provider as-is, resulting in a server error
on the backend.

The fix is to encode certain properties as base64, which makes them
perfectly safe for JSON parsing. For now, only the `metadata` property gets
such special treatment.

The result is that the AUTH_PROVIDERS value is even more unreadable than it
was before. Granted, someone wanting to reading bare XML is unlikely, but it
now makes the configuration file look even more intimidating than before.

npm test passes