fix: again use classic settings when possible
In a follow-up to the attempts to address a problem in which some
configuration values had characters that caused dotenv to raise an error,
this change will encode the one known property (SAML IdP metadata) using
base64. This was already being done when writing the providers to the
AUTH_PROVIDERS setting, now it is also applied to SAML_IDP_METADATA.

With this change, it is now feasible once again to prefer to use the classic
OIDC_* and SAML_* settings when possible, rather than always writing to the
AUTH_PROVIDERS setting. This will improve the readability of the .env file
and reduce the burden on support when interpreting customer configurations.

npm test passes