fix: no longer save empty authnContext to config
When adding a new SAML provider, the service would write an empty string for
the SAML_AUTHN_CONTEXT setting to the .env file. This led to the frontend
reporting an error (ctx is undefined). It should not be possible, or even
necessary, to save an empty value for this setting, as the only way to
disable the authnContext is to set the appropriate setting, which is
SAML_DISABLE_CONTEXT.

npm test passes