fix: report group names containing spaces as error
When provisioning a group whose name contains a space character, report that
as an error to the cloud service provider. Perforce group names cannot
contain spaces, and reporting the error properly to the CSP will help in
surfacing this to the admin. Previously the operation would simply fail
mysteriously with a misleading error message in the service log.

npm test passes