build: fix Linux package build process in docker
The `docker run` command to extract the build artifact started failing
intermittently, resulting in apparent package signing issues. Now the
process creates an image based on `scratch` that contains only the build
artifacts, and then we use the `--output` flag to `docker build` to write
the results to the local disk rather than the build image.