build: put admin content in private directory
Despite not enabling the administrative interface, the static content for
the admin interface was still being served to the client. This was because
the content was under the `public` directory and hence is served in the same
manner as all other public content. Now the build puts the admin files under
a directory named `private` that will prevent this from happening.