feat: verify client cert sent via HTTP header
If the service is running behind a reverse proxy that is both terminating
the TLS connection and forwarding the client certificate via an HTTP header,
then the service will read that cert and validate that either the subject
common name or the certifiate fingerprint matches the configuration.

This client cert validation is enabled by setting the new configuration
setting CLIENT_CERT_HEADER to the name of the HTTP header added by the
reverse proxy. This setting, along with either CLIENT_CERT_CN or
CLIENT_CERT_FP will be used to ensure the client certificate is the one
expected by the service.

Note that this is only used to validate clients requesting the user profile
data via either the /requests/status or /saml/validate endpoints.

npm test passes