fix: minimize reliance on process.env throughout
The delicate reliance on the process.env object in Node.js has been
relegated to a very small portion of the application. The "environment"
settings repository now reads the .env file and will return values from
that, falling back to process.env as necessary.

As part of the restructuring, the logger is now constructed in a usecase and
loaded by the dependency injection container, rather than being initialized
on startup by environment variables.

This clears the path for supporting alternative configuration file formats,
making it possible to configure the service will something sophisticated
enough to capture everything in a single file, such as TOML or YAML.

npm test passes