feat: allow optional config.toml for configuration
If the config.toml file exists at the time of startup, the service will use
that for configuration instead of the .env file. As before, environment
variables are still used if the configuration file lacks a particular
setting, and defaults.env is still used in either case. The difference is
that config.toml can capture everything that was previously stored in
separate files, including the logging, Swarm integration, Redis Sentinel
support, and multiple identity providers.

The use of the TOML naming convention (all lowercase with underscores
separating words) is optional.

npm test passes