feat: user provisioning for multiple servers
The user provisioning feature now supports multiplexing of the incoming
cloud service providers and the backend Helix Core Server instances. This is
configured using a new `PROVISIONING` setting that defines the allowed
domains and their associated servers. An incoming request is assigned to a
domain by their unique bearer token value. Each domain label must be
assigned by the administrator. One or more servers can be associated with
one or more domains. If multiple servers are associated with the same
domain, one of them must have that domain in their `leader` list. All reads
will be directed to the leader, while writes will go to the leader first,
followed by the other members in no particular order. Any failures result in
an immeidate return with partial writes left in place. Support for rollbacks
might be feasible in the future, if there is a genuine need.

If using the original BEARER_TOKEN and P4* settings, then there are no
changes to the behavior, provisioning will continue to work as before.

Also in this change is a fix for the externalId value associated with users.
Previously this value was stored without regards to the provisioning domain,
so if multiple cloud service providers were to attempt to assign the
externalId, each would clobber the previously assigned value. By using the
`domain` value assigned by the administrator, the externalId value is scoped
to that domain. Each CSP may set their own externalId and when retrieving
the user record, it will contain the appropriate value for externalId. As
with the overall change, if using the existing settings, no changes to
behavior will be observed.

npm test passes