chore: update to Node.js v20 almost everywhere
Node.js v18 has entered the maintenance phase of its lifecycle as of last
month. As such, almost everything in the HAS source tree is now moving to
v20. The configure script will accept either v16, v18, or v20 since those
versions will continue to work. The install.sh script will install v20 and
the Linux packages will include the v20 binary.

Exceptions to this are Amazon Linux 2 and CentOS 7 which continue to use v16
of Node.js. This is due to their lack of support for more recent releases of
Node.js and there is nothing that we can be do about this. Anyone still
using those old releases will need to use older releases of the service.

npm test passes