fix: decode SAML IdP metadata for backward compat
The usecases are no longer handling the encoding or decoding of certain
values as the configuration file "data source" is now handing that. However,
the previous version shipped with code in place, for the admin webui, that
encoded the SAML IdP metadata using base64 without a value prefix. This is
not compatible with the new format, per se, so the usecase for loading the
authentication providers needs to detect and handle these values if they
appear to be encoded.

npm test passes