doc: example of rate limiting using HAProxy
Rather than adding more settings and behavior to the service, it is easier
to leverage HAProxy for rate limiting and request size limits. HAProxy is
already in the test environment and now we have a working configuration that
shows how to limit the number of connections and requests to the backend.

Note that HAProxy already limits the request size to 16kb, which achieves
the other goal of protecting the backend from large requests.

npm test passes