feat: structured logging for user provisioning
Any SCIM operation that modifies data will be logged to a new logging file
using JSON formatted lines, with as much detail as reasonable. This is on by
default and configured in the LOGGING configuration file in the new `scim`
section, which has the same structure and features of the usual logging
functionality (it is simply nested within the `scim` property). All errors
are also captured in this new log file, named provisioning.log. As with the
normal log file, this log will be rotated to control disk usage. All of this
can be changed, just like the normal logging, or entirely disabled by
removing the `scim` section of the configuration file.

npm test passes