fix: use P4TICKETS and P4TRUST for provisioning
The service was not loading the values for P4TICKETS or P4TRUST explicitly
and as a result, the settings were ignored in the new user provisioning
code. With this change, the user provisioning code will use the PROVISIONING
setting, if available, or defer to the classic P4PORT and friends. The new
PROVISIONING setting also may include p4tickets and p4trust values as well.

npm test passes