fix: install script for CentOS working once again
The procedure for installing Node.js from the NodeSource packages has
changed once again, back to the shell script. The script is much simpler and
manages the intricaies of setting up the package signing key. Previously the
script had been retired in favor of the manual package procedure, but that
is gone now and the script is back.

Details: the NodeSource signing key for the RPM packages was wrong, the one
provided by their prerequisite package did not match the one used to sign
the actual node package. Open issues are still pending, but meanwhile they
have already switched back to using the script anyway.