fix: correctly set and return externalId for users
Finishing the change that started earlier in which the user's externalId is
saved and returned to different cloud service providers. That is, for each
of the configured providers, their externalId will be updated and returned
as appropriately. Previously, most user operations were ignoring the value
and not updating or returning it as defined by the SCIM protocol.

npm test passes