fix: handle LOGGING read/write in consistently
When reading and writing the configuration file, either .env or the
config.toml file, handle the LOGGING setting consistently. If it was a file
path, then always read/write to that file; if it was the encoded or embedded
configuration (a la TOML), then always read/write it in that form.

npm test passes