feat: JS script to replace exposed certificates
The self-signed certificates that are included with the authentication
service are found on the internet, which makes them basically worthless for
anything other than sanity-checking an installation. This script serves to
corret that by providing a single command invocation to replace all of the
certs with new (self-signed) certs all around. Works on Linux and Windows.