feat: add /liveness HTTP endpoint for orchestration
Container orchestration systems like Kubernetes can benefit from having a
means of testing the readiness and/or liveness of the service. This is
especially meaningful when deploying the service with Redis -- when Redis is
configured but the connection is not working, the service will largely be
unable to fulfill requests.

The new endpoint (GET /liveness) is always accessible and returns a 200 if
everything is okay (either Redis is not configured, or the Redis connection
is working), or a 500 if Redis is configured but the connection is failing.

npm test passes