fix: treat routes/saml_idp.conf.cjs as config file
The APT and YUM packages were not treating the default IdP config file as a
configuration file, which resulted in any edits the user had made being
discarded when the package was updated. This change fixes that so the file
routes/saml_idp.conf.cjs will be treated like any other configuration file
(not removed when the package is removed or upgrade).