fix: ignore EOL when comparing Swarm, service certs
It is quite possible that the public key certificate on either the auth
service side or on the Swarm side will have differing end-of-line
characters. Previously, this would cause the Swarm config validator to
report an error since the values did not literally match byte for byte.
However, the EOL characters should not matter, and what's more, the Swarm
configuration can have a single line value for the x509cert setting. In
either case, the validator should accept the values as long as they have
the same base64 encoded value, regardless of EOL and BEGIN/END armor.

npm test passes