test: switch redis for valkey docker image
Now that Redis has changed to a strange license, it is safer to replace the
docker container with the Linux Foundation's fork, Valkey. Everything else
remains the same: configuration, documentation, tests, etc. The only change
is that the underlying image is now the Valkey image rather than Redis.

What should customers do about the Redis change? Probably nothing, but I am
not a lawyer. The change most likely only affects cloud service providers,
but the safest option is to switch to a compatible replacement, either
Valkey or KeyDB.

npm test passes