build: fix or add license info for Linux packages
The RPM package license was incorrectly specified as BSD (should be MIT),
and it was lacking the license file specifier. This additional field on the
RPM will indicate to the user where the full license text can be found.

Similarly, the APT package was missing the copyright text -- the file was
present but completely empty. It now contains the appropriately formatted
copyright and license information.