build: remove remaining Node.js v16 support
This affects CentOS/RHEL 7, support for which was already removed in a
previous change, as well as Amazon Linux 2. Both of these OS releases are so
old that Node.js v18 and later do not support them. Note that even v18 is
already in maintenance mode. The auth service needs to keep up with Node.js
as many of our dependencies also march forward to the future.