build: treat cert files as config files in packages
The Debian and RPM packages will now treat the certificate files (that are
included in the package) within the certs directory as if they are
configuration files. This means that on package remove or upgrade,
the (modified) files will be left in place on remove and not overwritten by
an upgrade.