fix: SCIM user/group update now merges spec values
When updating a user or group via SCIM the existing fields in the spec that
are not effected by SCIM were being reset. With this change, the user and
group update operations will output the existing spec and merge that with
the values coming from the cloud service provider.

For isntance, the group field "Timeout" would get reset to the default value
whenever the group is updated.

npm test passes