fix: add LOGGING to .env file on install
Previously the install script was simply creating the .env file if it was
missing, by copying the example.env and stripping comments and blank lines.
However, the package version also injects a LOGGING line, so this change
brings parity to the package experience by doing the same.

This is done only if the .env and config.toml files are missing. Also note
that previously the .env file was created only if the systemd service unit
was being created, which didn't make sense, so it is now separate from that
task, which is optional thanks to the --no-systemd flag.