feat: treat BEARER_TOKEN as file if it is one
If the BEARER_TOKEN setting is not a plain string value but instead names a
file, then log a warning and read the token from the file. Ideally the token
should be stored in a file that is named by BEARER_TOKEN_FILE, not the old
setting. Nonetheless, in an effort to make the service do what would seem
most sensible, simply read the file if BEARER_TOKEN refers to a file.

npm test passes