fix: groom cert and key file for SAML
When reading the public/private key files for preparing a SAML login
request, ensure the content only has the encapsulated base64 encoded data
and the BEGIN/END text armor. Any preceding attribute values are stripped,
as that causes a problem with the node-saml library, which is seemingly not
compliant with the recommendations described in RFC 7468.

npm test passes