fix: groom cert now includes all BEGIN/END blocks
Previously the fix for the node-saml issue (attributes that preceded the
encapsulation boundary in the PEM encoded certificate would result in an
error when authenticating using SAML) was only emitting the first block
wrapped in the BEGIN/END boundary lines. However, that is not correct for
files that contain a CA _and_ the server certificate. Now the grooming will
emit all BEGIN/END blocks and elide anything else.

npm test passes