fix: configure script tolerates running as root
The configure shell script that ensures the system is ready before
attempting to run the JavaScript version of the configure script is now
tolerant of running as the root user. Previously, if the script was run as
the root user, it would subsequently be unable to run as a non-root user.