fix: retain Owners property when patching groups
When the user provisioning feature updates groups in Perforce, it was losing
certain fields such as Owners. This would happen whenever users are added or
removed from the group. While this has no behavioral impact, it undoes the
fix for keeping a group around when there are no members.

npm test passes