fix: webui, do not assume authnContext is an array
The web interface was assuming that the authnContext property of an auth
provider was always an array. That was probably true at some point in the
past but that has changed such that it may be a) undefined, b) a single
string value, or c) an array of string values.