fix: do not inject default values erroneously
If any of the auth providers did not have a `protocol` then it was treated
as SAML and all of the default SAML settings were being inserted into the
provider configuration. No harm, but it does create confusion for anyone
looking at the results. This was surfaced while investigating the issue
concerning the authnContext error in the admin interface.

npm test passes