fix: honor OIDC_MAX_AGE properly during login
Setting OIDC_MAX_AGE to zero is supposed to force the user to authenticate
every time (support for this varies among identity providers, however). The
code was using the wrong logic to check for this configurable such that if
it was zero, it had no effect. A value of 1 would work effectively the same,
but the documentation suggests setting to zero should work.

npm test passes