test: add --no-ping option to install script
The VM cluster for automated builds has unreliable network connectivity and
as such the tests often fail at the step in which the install script pings
ubunut.com to ensure the apt commands will succeed. This new option disables
that ping in the hopes that the tests might succeed more often than not.