fix: ignore prefix when updating user via SCIM
A SCIM request to update a user may have the user identifier in the URL with
the `user-` prefix that P4AS likes to add to user entities. Unfortunately,
the code for updating a user would see this as an attempt to rename the user
and fail because renames are not allowed by default.

npm test passes