fix: provisioning log indicates if group changed
Previously the provisioning log would always indicate that a group was
updated every time a PATCH or PUT was received, even if nothing changed in
the group within P4 Server. With this change the service will now clearly
indicate if a group changed or not in the logs.

npm test passes