feat: allow outbound proxy for OIDC connections
In theory this change, according to the customer, will allow outbound OIDC
connections to go through a proxy. This is handled in the proxy-agent 3rd
party module and configured with the http_proxy, https_proxy, and no_proxy
environment variables that are well-known.

SAML is not supported because passport-saml does not support proxies nor
does it leverage an http agent of any sort. While the global-agent 3rd party
module might work, it has not been updated in several years.

Not documented

Not tested

npm test passes