fix: check output of openssl when validating certs
Seems that openssl command has changed again and now the exit code is always
zero, even if something might be wrong with the inputs. Now the service will
search for "will expire" in the output of the openssl command in order to
determine if the certificates being tested will expire soon.

npm test passes