fix: trim group members when logging for SCIM
When logging SCIM operations on groups, in particular the POST and PUT, trim
the members list to at most 10. Sometimes a very large group is updated and
the entire long list of members would get written to the provisioning log.
This is almost always unnecessary and if a group is updated 1000 times to
add 1000 members one at a time, the log will group very large.

npm test passes