fix: do not add P4 settings unless required
When running the configure script, the P4 settings should not be added to
the .env file unless the admin provides those values non-interactively, or
chooses to configure the SCIM user provisioning interactively.

Previously, of the p4 binary was in the path and P4PORT and P4USER were
defined, then they would be added to the .env file unconditionally.

If the settings are already in the .env file, they are left intact.