feat: allow hiding the home page
As with the /status route, it is now possible to configure the service to
return a "not found' when the home page (/) is requested. By default the
home page will be enabled. By adding HOME_PAGE_ENABLED=false to the .env
file the / route will return a 404 and show the "not found" error page.

npm test passes