fix: add a retry to server requests
There was a timing issue when the service would perform a soft restart after
the admin interface had made changes to the configuration. It could happen
that during the admin interface refresh, the request to the service would
fail and result in an error being displayed in the browser.

This change adds a retry in the client code to simply try the same request
again after a short delay, with a maximum of 8 retries. This prevents the
error from being surfaced in the browser since it is not an error that the
admin needs to do anything about.