fix: add IdP base URL to CSP form-action list
When PROMPT_FOR_AUTHORIZATION is set in the configuration and the user is
using Google Chrome to authenticate, the "Approve Login" button would appear
to do nothing. If the browser console is opened, it shows that Chrome is
rejecting the request because the page is attempting to post form data to a
different URL (the identity provider) and that violates the content security
policy that helmet.js produces by default.

This change adds all of the identity provider base URLs to the list of
values in the 'form-action' property of the content security policy. For an
OIDC provider, the issuer URI will be used, while for SAML it will be the
IdP SSO URL (the service will fetch and parse the metadata if necessary).

npm test passes