test: establish docker container dependencies
The docker containers now have an established order of dependency, such that
one container is not considered ready to start until another one is healthy.
With this, the auth service containers will wait for Shibboleth to be ready
before starting, which avoids the SAML metadata fetching error. This type of
error should never occur in production -- no real IdP is expected to be down
when the auth service is starting. This is a test environment problem.